Recursively clean up conflicting animations
Summary:
Call getAndEraseConflictingAnimations recursively, to clean up any conflicting animations.

With a specific sequence of mutation instructions, it is possible that an animation on a tree is set up, and the entire tree needs to be cleaned up if a parent is deleted, say.

This can happen especially if rapidly mutating trees in such a way that invokes view flattening and unflattening (which is not recommended, but is certainly possible).

Changelog: [Internal]

Reviewed By: sammy-SC

Differential Revision: D27407776

fbshipit-source-id: 5125250f051c58870692f8fdc43ed23da5058a7f