add release changelog back to core repo (#32520)
Summary:
As agreed, we are moving back the changelog from the RNC repo into the core one. This PR does just that.

It will be followed up by a new draft PR for the 0.67 release ([old one](https://github.com/react-native-community/releases/pull/255)).

Once this is merged, a few links will need to be fixed, such as:
* https://github.com/facebook/react-native/wiki/Changelog
* https://github.com/facebook/react-native/releases (for new releases)

Moreover, we should have a conversation about how changes are reported for the "monorepo" part of this repo: https://github.com/facebook/react-native/tree/main/packages as far as I can see, the changes to those packages are not documented anywhere.

## Changelog

[Internal] [Added] - Adding Changelog.md back into main repo

Pull Request resolved: https://github.com/facebook/react-native/pull/32520

Test Plan: N/A

Reviewed By: lunaleaps

Differential Revision: D32102055

Pulled By: cortinico

fbshipit-source-id: 6d1f353bb6acdc952e40941510ca36200690bbf8